backend/requirements.txt
Django==3.2.16
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
gunicorn==20.1.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
idna==3.7
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.1
Pillow
pluggy==0.13.1
py==1.11.0
pycparser==2.22
PyJWT==2.9.0
asgiref==3.8.1
atomicwrites==1.4.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==43.0.1
defusedxml==0.8.0rc2
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.26.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
toml==0.10.2
typing_extensions==4.12.2
urllib3==1.26.19
psycopg2-binary
flake8==6.0.0
flake8-isort==6.0.0
numpy